Upgrade the Neurai blockbook from the upstream snapshot 674271fe
(2025-10-15, "0.5.0") to trezor/blockbook v0.6.0.

Neurai-specific adaptations kept/applied on top of upstream:
- bchain/coins/neurai (parser, rpc, tests) unchanged: it mirrors the
  Ravencoin implementation, whose upstream API did not change.
- bchain/coins/blockchain.go: keep "Neurai" / "Neurai HW" registration.
- configs/coins/neurai.json: add rpc_url_ws_template (new in 0.6.0).
- docs/ports.md: register Neurai ports (9169/9069/8069/38369).
- Makefile / build/docker/bin/Dockerfile: keep BASE_IMAGE=debian:12,
  take upstream's pinned DOCKER_VERSION/DOCKER_SHA256.
- Remove leftovers already deleted upstream (goerli/ropsten configs and
  testdata, bitcoin-signet.conf, static/*.min.3.*).

+++ bchain/coins/btc/bitcoinparser.go
@@ -80,9 +80,14 @@ type Vout struct {
 // Tx is blockchain transaction
 // unnecessary fields are commented out to avoid overhead
 type Tx struct {
-	Hex         string       `json:"hex"`
-	Txid        string       `json:"txid"`
-	Version     int32        `json:"version"`
+	Hex  string `json:"hex"`
+	Txid string `json:"txid"`
+	// Version is decoded as uint32 to tolerate non-standard/invalid tx
+	// versions present on Bitcoin mainnet (e.g. tx 637dd1a3...fef7413f in
+	// block 256818 has version 2187681472, which overflows int32). It is
+	// bit-cast to int32 in ParseTxFromJson to match the value the binary
+	// block parser produces via wire.MsgTx.Deserialize.
+	Version     uint32       `json:"version"`
 	LockTime    uint32       `json:"locktime"`
 	VSize       int64        `json:"vsize,omitempty"`
 	Vin         []bchain.Vin `json:"vin"`
@@ -108,7 +113,7 @@ func (p *BitcoinParser) ParseTxFromJson(msg json.RawMessage) (*bchain.Tx, error)
 	// it is necessary to copy bitcoinTx to Tx to make it compatible
 	tx.Hex = bitcoinTx.Hex
 	tx.Txid = bitcoinTx.Txid
-	tx.Version = bitcoinTx.Version
+	tx.Version = int32(bitcoinTx.Version)
 	tx.LockTime = bitcoinTx.LockTime
 	tx.VSize = bitcoinTx.VSize
 	tx.Vin = bitcoinTx.Vin
